moved kbn-truncated to notify (#33374)
