+++ src/legacy/ui/public/doc_table/doc_table.js
@@ -20,7 +20,6 @@
 import _ from 'lodash';
 import html from './doc_table.html';
 import { getSort } from './lib/get_sort';
-import '../directives/truncated';
 import '../directives/infinite_scroll';
 import './components/table_header';
 import './components/table_row';
