+++ src/legacy/ui/public/notify/directives/truncated.js
@@ -19,8 +19,8 @@
 
 import truncText from 'trunc-text';
 import truncHTML from 'trunc-html';
-import { uiModules } from '../modules';
-import truncatedTemplate from './partials/truncated.html';
+import { uiModules } from '../../modules';
+import truncatedTemplate from '../partials/truncated.html';
 import 'angular-sanitize';
 
 const module = uiModules.get('kibana', ['ngSanitize']);
