+++ src/legacy/ui/public/notify/notify.js
@@ -26,7 +26,7 @@ import { banners } from './banners';
 import { Notifier } from './notifier';
 import template from './partials/toaster.html';
 import '../filters/markdown';
-import '../directives/truncated';
+import './directives/truncated';
 import { FormattedMessage } from '@kbn/i18n/react';
 
 import {
