+++ src/legacy/ui/public/notify/partials/truncated.html
